+++ src/pages/Home/component/BusState/component/BusStateCard/styles.ts
